+++ apps/bfd-pipeline/bfd-pipeline-rda-grpc/src/test/java/gov/cms/bfd/pipeline/rda/grpc/sink/direct/McsClaimRdaSinkTest.java
@@ -236,7 +236,7 @@ public class McsClaimRdaSinkTest {
             RdaChange.Type.UPDATE,
             claim,
             changeDate,
-            new RdaChange.Source("p1", 0, "1970-01-01T00:00:00.000000Z"));
+            new RdaChange.Source("p1", (short) 0, "1970-01-01T00:00:00.000000Z"));
     RdaClaimMessageMetaData metaData = sink.createMetaData(change);
     assertEquals(100L, metaData.getSequenceNumber());
     assertEquals('M', metaData.getClaimType());
@@ -271,6 +271,6 @@ public class McsClaimRdaSinkTest {
         RdaChange.Type.INSERT,
         claim,
         clock.instant().minusMillis(12),
-        new RdaChange.Source("p1", 0, "1970-01-01T00:00:00.000000Z"));
+        new RdaChange.Source("p1", (short) 0, "1970-01-01T00:00:00.000000Z"));
   }
 }
